pyproject.toml
[project]
name = "eden-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.8"
dependencies = [
    "django>=5.2.6",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.28.0",
    "drf-spectacular-sidecar>=2025.10.1",
    "python-decouple>=3.8",
]
